+++ lib/screens/signal/signal_list.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:periodik/providers/points_provider.dart';
+import 'package:periodik/screens/point/point_dialog.dart';
 
 class SignalList extends ConsumerWidget {
   const SignalList({
@@ -17,7 +18,19 @@ class SignalList extends ConsumerWidget {
       itemBuilder: (context, index) {
         final point = points[index];
         return ListTile(
-          onTap: () {},
+          onTap: () {
+            PointDialog.show(
+              context: context,
+              signalId: id,
+              point: point,
+            );
+          },
+          trailing: point.state
+              ? Icon(
+                  Icons.circle,
+                  color: Theme.of(context).colorScheme.error,
+                )
+              : null,
           title: Text(point.date.toString()),
           subtitle: Text(point.state.toString()),
         );