+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## UNRELEASED
+
+- Allows the edition of a point.
+- Makes Monday the first day of the week in the calendar view.
+
 ## 0.8.0
 
 - Allows adding a new point on clicking on an empty calendar day.